requirements.txt
streamlit==1.35.0
pandas==2.2.2
python-dotenv==1.0.1
anthropic==0.28.0
plotly==5.22.0
openpyxl==3.1.4
